// src/app/models/profile.model.ts
export interface Profile {
  address: string;
  bank_number: string;
  birth: Date;
  city: string;
  city_id: string;
  count_order_success: number;
  country: string;
  created_at: string;
  district_id: string;
  door: string;
  email: string;
  email1: string;
  email_apple: string;
  email_google: string;
  extra_info: string;
  fb_id: string;
  fcm_token: string;
  firstName: string;
  first_time_make_order: string;
  first_time_open_app: string;
  flat: string;
  full_name: string;
  gender: string;
  id: number;
  is_apartment: string;
  is_block: string;
  is_influencer: number;
  lastName: string;
  password: string;
  phone_number: string;
  phone_number_firebase: string;
  real_name: string;
  state: string;
  street_number: string;
  token: string;
  town: string;
  type: string;
  updated_at: string;
  username: string;
  ward_id: string;
  zip: string;
}